[auth] Move jetauth logic to /pkg
