+++ envsec/pkg/envcli/auth.go
@@ -4,15 +4,12 @@
 package envcli
 
 import (
-	"context"
 	"fmt"
-	"time"
 
 	"github.com/spf13/cobra"
-	"go.jetpack.io/envsec/internal/envvar"
-	"go.jetpack.io/envsec/internal/jetcloud"
 	"go.jetpack.io/pkg/sandbox/auth"
-	"go.jetpack.io/pkg/sandbox/auth/session"
+	"go.jetpack.io/pkg/sandbox/envvar"
+	"go.jetpack.io/pkg/sandbox/jetauth"
 )
 
 func authCmd() *cobra.Command {
@@ -135,32 +132,5 @@ func newAuthClient() (*auth.Client, error) {
 	// TODO: Consider making scopes and audience configurable:
 	// "ENVSEC_AUTH_SCOPE" = "openid offline_access email profile"
 	// "ENVSEC_AUTH_AUDIENCE" = "https://api.jetpack.io",
-	return auth.NewClient(
-		issuer,
-		clientID,
-		getShortTermAccessToken,
-		getShortTermAccessToken,
-	)
-}
-
-func getShortTermAccessToken(
-	ctx context.Context,
-	tok *session.Token,
-) (*session.Token, error) {
-	accessToken, err := jetcloud.GetAccessToken(ctx, tok)
-	if err != nil {
-		fmt.Println(err)
-		// We set the current set of tokens to be expired immediately.
-		// This is a hack to force a refresh. Even though we have a new id token,
-		// we failed to get a valid access token. This is likely because the
-		// user doesn't have a valid plan.
-		// Returning a nil token would prevent any new data from being written to
-		// token store, but that means our existing refresh token would no longer
-		// be valid because it is only usable once.
-		tok.Expiry = time.Now()
-		tok.AccessToken = ""
-	} else {
-		tok.AccessToken = accessToken
-	}
-	return tok, nil
+	return jetauth.NewClient(issuer, clientID)
 }
