+++ envsec/pkg/envcli/flags.go
@@ -10,10 +10,10 @@ import (
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 	"go.jetpack.io/envsec"
-	"go.jetpack.io/envsec/internal/jetcloud"
-	"go.jetpack.io/envsec/internal/typeids"
 	"go.jetpack.io/envsec/pkg/awsfed"
 	"go.jetpack.io/pkg/sandbox/auth/session"
+	"go.jetpack.io/pkg/sandbox/jetcloud"
+	"go.jetpack.io/pkg/sandbox/typeids"
 )
 
 // to be composed into xyzCmdFlags structs
